use component melee knockback
Vanilla computes melee knockback from the attacker's ATTACK_KNOCKBACK attribute and then applies the attacking weapon's KNOCKBACK enchantment value effect. The local helper still returned only the old integer Knockback enchantment level from the current main hand. That dropped attribute knockback, fractional or custom component-driven knockback values, and source-stack correctness for offhand spear paths.

The fix changes the enchantment knockback helper to return a double based on ATTACK_KNOCKBACK plus the weapon's KNOCKBACK value effect. Normal melee keeps using the main-hand stack, while spear piercing and kinetic stab paths pass their actual spear stack so offhand and custom weapons use the same source item vanilla would use.

+++ src/main/java/io/github/togar2/pvp/feature/attack/AttackValues.java
@@ -2,7 +2,7 @@ package io.github.togar2.pvp.feature.attack;
 
 public class AttackValues {
 	public record PreCritical(float damage, float magicalDamage, double cooldownProgress,
-	                          boolean strong, boolean sprint, int knockback, int fireAspect) {
+	                          boolean strong, boolean sprint, double knockback, int fireAspect) {
 		public PreSweeping withCritical(boolean critical) {
 			return new PreSweeping(
                     this.damage, this.magicalDamage, this.cooldownProgress,
@@ -13,7 +13,7 @@ public class AttackValues {
 
 	public record PreSweeping(float damage, float magicalDamage, double cooldownProgress,
 	                          boolean strong, boolean sprint, boolean critical,
-	                          int knockback, int fireAspect) {
+	                          double knockback, int fireAspect) {
 		public PreSounds withSweeping(boolean sweeping) {
 			return new PreSounds(
                     this.damage, this.magicalDamage, this.cooldownProgress,
@@ -25,10 +25,10 @@ public class AttackValues {
 
 	public record PreSounds(float damage, float magicalDamage, double cooldownProgress,
 	                  boolean strong, boolean sprint, boolean critical, boolean sweeping,
-	                  int knockback, int fireAspect) {}
+	                  double knockback, int fireAspect) {}
 
 	public record Final(float damage, float baseDamage, double cooldownProgress, boolean strong, boolean sprint,
-	                    int knockback, boolean critical, boolean magical,
+	                    double knockback, boolean critical, boolean magical,
 	                    int fireAspect, boolean sweeping, boolean sounds,
 	                    boolean playSoundsOnFail) {}
 }
